Assert: Activation
