+++ test/Assert.h
@@ -1,0 +1,8 @@
+
+#ifndef FL_TEST_ASSERT_H
+#define FL_TEST_ASSERT_H
+
+#include "test/AssertActivation.h"
+#include "test/AssertConstructor.h"
+
+#endif
