+++ test/Headers.h
@@ -4,7 +4,6 @@
 
 #include <catch2/catch_all.hpp>
 
-#include "AssertConstructor.h"
 #include "fuzzylite/Headers.h"
 
 namespace fuzzylite { namespace test {
