+++ test/TestTerm.cpp
@@ -13,7 +13,8 @@
 
  fuzzylite is a registered trademark of FuzzyLite Limited.
  */
-#include "Headers.h"
+#include "test/Assert.h"
+#include "test/Headers.h"
 
 namespace fuzzylite { namespace test {
 
@@ -283,8 +284,10 @@ namespace fuzzylite { namespace test {
         const AlgebraicProduct algebraicProduct;
         Triangle term("triangle", -0.400, 0.000, 0.400);
         TermAssert<Activated>(&term, 1.0, &algebraicProduct)
-            .repr_is("fl.Activated(term=fl.Triangle('triangle', -0.4, 0.0, 0.4), "
-                     "degree=1.0, implication=fl.AlgebraicProduct())")
+            .repr_is(
+                "fl.Activated(term=fl.Triangle('triangle', -0.4, 0.0, 0.4), "
+                "degree=1.0, implication=fl.AlgebraicProduct())"
+            )
             .exports_fll("AlgebraicProduct(1.000,triangle)", false)
             .is_not_monotonic()
             .has_memberships({
@@ -305,8 +308,10 @@ namespace fuzzylite { namespace test {
             });
 
         TermAssert<Activated>(&term, 0.5, &algebraicProduct)
-            .repr_is("fl.Activated(term=fl.Triangle('triangle', -0.4, 0.0, 0.4), "
-                     "degree=0.5, implication=fl.AlgebraicProduct())")
+            .repr_is(
+                "fl.Activated(term=fl.Triangle('triangle', -0.4, 0.0, 0.4), "
+                "degree=0.5, implication=fl.AlgebraicProduct())"
+            )
             .exports_fll("AlgebraicProduct(0.500,triangle)", false)
             .configured_as("###")
             .exports_fll("AlgebraicProduct(0.500,triangle)");
@@ -375,11 +380,13 @@ namespace fuzzylite { namespace test {
 
         TermAssert<Aggregated>(aggregated)
             .exports_fll("fuzzy_output: Aggregated Maximum[Minimum(0.600,LOW),Minimum(0.400,MEDIUM)]", false)
-            .repr_is("fl.Aggregated(name='fuzzy_output', minimum=-1.0, maximum=1.0, "
-                     "aggregation=fl.Maximum(), terms=[fl.Activated(term=fl.Triangle('LOW', -1.0, "
-                     "-0.5, 0.0), degree=0.6, implication=fl.Minimum()), "
-                     "fl.Activated(term=fl.Triangle('MEDIUM', -0.5, 0.0, 0.5), degree=0.4, "
-                     "implication=fl.Minimum())])")
+            .repr_is(
+                "fl.Aggregated(name='fuzzy_output', minimum=-1.0, maximum=1.0, "
+                "aggregation=fl.Maximum(), terms=[fl.Activated(term=fl.Triangle('LOW', -1.0, "
+                "-0.5, 0.0), degree=0.6, implication=fl.Minimum()), "
+                "fl.Activated(term=fl.Triangle('MEDIUM', -0.5, 0.0, 0.5), degree=0.4, "
+                "implication=fl.Minimum())])"
+            )
             .is_not_monotonic()
             .has_memberships(
                 {
@@ -766,8 +773,10 @@ namespace fuzzylite { namespace test {
             .exports_fll("term: discrete Discrete 0.000 1.000 8.000 9.000 4.000 5.000 2.000 3.000 6.000 7.000")
             .apply([](Term* term) -> void { dynamic_cast<Discrete*>(term)->sort(); })
             .exports_fll("term: discrete Discrete 0.000 1.000 2.000 3.000 4.000 5.000 6.000 7.000 8.000 9.000")
-            .repr_is("fl.Discrete('discrete', fl.array([fl.array([0.0, 1.0]), fl.array([2.0, "
-                     "3.0]), fl.array([4.0, 5.0]), fl.array([6.0, 7.0]), fl.array([8.0, 9.0])]))");
+            .repr_is(
+                "fl.Discrete('discrete', fl.array([fl.array([0.0, 1.0]), fl.array([2.0, "
+                "3.0]), fl.array([4.0, 5.0]), fl.array([6.0, 7.0]), fl.array([8.0, 9.0])]))"
+            );
 
         CHECK_THROWS_WITH(
             Discrete("discrete").membership(0), Catch::Matchers::StartsWith("[discrete error] term is empty")
@@ -776,14 +785,20 @@ namespace fuzzylite { namespace test {
         TermAssert<Discrete>("discrete")
             .configured_as("0 1 8 9 4 5 2 3 6 7 0.5")
             .apply([](Term* term) -> void { dynamic_cast<Discrete*>(term)->sort(); })
-            .exports_fll("term: discrete Discrete "
-                         "0.000 1.000 2.000 3.000 4.000 5.000 6.000 7.000 8.000 9.000 0.500")
-            .repr_is("fl.Discrete('discrete', fl.array([fl.array([0.0, 1.0]), fl.array([2.0, "
-                     "3.0]), fl.array([4.0, 5.0]), fl.array([6.0, 7.0]), fl.array([8.0, 9.0])]), "
-                     "0.5)")
+            .exports_fll(
+                "term: discrete Discrete "
+                "0.000 1.000 2.000 3.000 4.000 5.000 6.000 7.000 8.000 9.000 0.500"
+            )
+            .repr_is(
+                "fl.Discrete('discrete', fl.array([fl.array([0.0, 1.0]), fl.array([2.0, "
+                "3.0]), fl.array([4.0, 5.0]), fl.array([6.0, 7.0]), fl.array([8.0, 9.0])]), "
+                "0.5)"
+            )
             .configured_as(" -0.500 0.000 -0.250 1.000 0.000 0.500 0.250 1.000 0.500 0.000")
-            .exports_fll("term: discrete Discrete "
-                         "-0.500 0.000 -0.250 1.000 0.000 0.500 0.250 1.000 0.500 0.000")
+            .exports_fll(
+                "term: discrete Discrete "
+                "-0.500 0.000 -0.250 1.000 0.000 0.500 0.250 1.000 0.500 0.000"
+            )
             .has_memberships({
                 {-1.0, 0.0},
                 {-0.5, 0.0},
@@ -2199,7 +2214,8 @@ namespace fuzzylite { namespace test {
             .infix_is("3.000 ^ 4.000")
             .evaluates_to(81.0);
 
-        std::unique_ptr<Function::Node> node_sin(new Function::Node(ff.cloneObject("sin"), fl::null, node_pow->clone())
+        std::unique_ptr<Function::Node> node_sin(
+            new Function::Node(ff.cloneObject("sin"), fl::null, node_pow->clone())
         );
         FunctionNodeAssert(node_sin->clone())
             .postfix_is("3.000 4.000 ^ sin")
@@ -2212,8 +2228,10 @@ namespace fuzzylite { namespace test {
             .postfix_is("3.000 4.000 ^ sin two pow")
             .prefix_is("pow sin ^ 3.000 4.000 two")
             .infix_is("pow ( sin ( 3.000 ^ 4.000 ) two )")
-            .fails_to_evaluate("[function error] expected a map of variables containing the value for 'two', "
-                               "but none was provided")
+            .fails_to_evaluate(
+                "[function error] expected a map of variables containing the value for 'two', "
+                "but none was provided"
+            )
             .evaluates_to(0.39675888533109455, {{"two", 2}});
 
         std::unique_ptr<Function::Node> node_sum(
